Moved admin controllers into account/admin namespace, fixed appropriate tests.
