+++ app/views/account/admin/issues/index.html.slim
@@ -1,5 +1,4 @@
 h1 Issues
-= link_to 'Home', root_path
 .container
   .table-responsive
     table.table.table-bordered
