+++ app/views/account/admin/users/_user.html.slim
@@ -4,12 +4,12 @@ tr
   td
     - if policy(user).toggle_ban?
       - if user.banned?
-        = link_to 'Unban', toggle_ban_account_user_path(user), method: :patch
+        = link_to 'Unban', toggle_ban_account_admin_user_path(user), method: :patch
       - else
-        = link_to 'Ban', toggle_ban_account_user_path(user), method: :patch
+        = link_to 'Ban', toggle_ban_account_admin_user_path(user), method: :patch
   td
     - if policy(user).change_role?
-      = simple_form_for(user, url: change_role_account_user_path(user)) do |f|
+      = simple_form_for(user, url: change_role_account_admin_user_path(user)) do |f|
         = f.select(:role, User.roles.keys)
         button.btn.btn-default.btn-xs = f.submit 'Change role'
     - else
