+++ app/views/account/admin/users/index.html.slim
@@ -1,5 +1,4 @@
 h1 All Users
-= link_to 'Home', root_path
 .container
   .table-responsive
     table.table.table-bordered
