+++ spec/controllers/account/admin/admin_users_controller_spec.rb
@@ -1,20 +1,24 @@
 require 'rails_helper'
 
-describe Account::UsersController, type: :controller do
+describe Account::Admin::UsersController, type: :controller do
   let(:reporter) { create(:user, :reporter) }
   let(:admin) { create(:user, :admin) }
   let!(:moderator) { create(:user, :moderator) }
 
   describe 'GET #index' do
+    before :each do
+      @request.env['HTTP_REFERER'] = account_admin_issues_url
+    end
+    
     it 'when user is not logged in' do
       get :index
-      expect(response).not_to redirect_to account_users_path
+      expect(response).not_to redirect_to account_admin_users_path
     end
 
     it 'when user is a reporter' do
       sign_in reporter
       get :index
-      expect(response).not_to redirect_to account_users_path
+      expect(response).not_to redirect_to account_admin_users_path
     end
 
     it 'when user is a moderator' do
@@ -31,6 +35,10 @@ describe Account::UsersController, type: :controller do
   end
 
   describe 'PATCH #change_role' do
+    before :each do
+      @request.env['HTTP_REFERER'] = account_admin_issues_url
+    end
+    
     it 'when user is not logged in' do
       patch :change_role, params: { id: moderator.id, user: { role: 'reporter' } }
       moderator.reload
